Split the Altimetry god-class into source + plotting modules (#130)
Break the 3807-line `altimetry.py` (one `Altimetry` class, ~50 methods,
four unrelated responsibilities) into single-concern modules behind the
existing test net, with `Altimetry` reduced to a thin coordinator.

New modules:
- `icesat2_source.py` — `Icesat2Source`: SlideRule fetch + parquet caching,
  ESA WorldCover sampling, temporal/outlier filtering, DEM differencing,
  and best-track / best-worst-segment selection.
- `planetary_source.py` — `PlanetarySource`: LOLA/MOLA CSV loaders + DEM
  differencing; absorbs the module-level `gds_query_async` / `GDS_BASE_URL`
  (re-exported from `altimetry` so the CLI import is unchanged).
- `altimetry_plots.py` — `AltimetryPlotter`: all figure rendering, operating
  on already-prepared dataframes.

`Altimetry` now composes the three objects (each holds a back reference to
the coordinator, which owns the cross-cutting `directory` / `dem_fn` /
`aligned_dem_fn`) and keeps the pc_align orchestration + shared keep/discard
decision (#127). The full public/notebook API is preserved by delegation, so
`report_pipeline.py` and notebooks are unchanged; `Altimetry`,
`AlignmentResult`, and `ICESAT2_MISSION_START` remain importable from
`asp_plot.altimetry`.

Plotting no longer triggers heavy I/O mid-render: the coordinator computes
the dh columns (`atl06sr_to_dem_dh` / `planetary_to_dem_dh`) and resolves the
track *before* delegating, so plot bodies receive prepared dataframes (only a
hillshade-backdrop read remains, which is a rendering concern). The over-long
`plot_atl06sr_dem_profile` and `plot_best_worst_segments` are decomposed into
panel helpers; the shared `nmad` helper moves to `utils`.

No single module exceeds ~1400 LOC. White-box tests are re-pointed at the new
objects (`alt.icesat2` / `alt.planetary`) and a composition-structure test is
added. Also fixes a latent shared-mutable-default-arg footgun in the
constructor (the ATL06-SR dicts are now fresh per instance).

Closes #130
